+++ Source/WTF/wtf/WeakRef.h
@@ -25,9 +25,16 @@
 
 #pragma once
 
-#include <wtf/WeakPtr.h>
+#include <wtf/GetPtr.h>
+#include <wtf/HashTraits.h>
+#include <wtf/SingleThreadIntegralWrapper.h>
+#include <wtf/ThreadSafeRefCounted.h>
+#include <wtf/Threading.h>
+#include <wtf/TypeCasts.h>
 
 namespace WTF {
+
+enum class EnableWeakPtrThreadingAssertions : bool { No, Yes };
 
 // Similar to a WeakPtr but it is an error for it to become null. It is useful for hardening when replacing
 // things like `Foo& m_foo`. It is similar to CheckedRef but it generates crashes that are more actionable.
@@ -44,11 +51,22 @@ public:
         UNUSED_PARAM(shouldEnableAssertions);
     }
 
+    explicit WeakRef(Ref<WeakPtrImpl>&& impl, EnableWeakPtrThreadingAssertions shouldEnableAssertions = EnableWeakPtrThreadingAssertions::Yes)
+        : m_impl(WTFMove(impl))
+#if ASSERT_ENABLED
+        , m_shouldEnableAssertions(shouldEnableAssertions == EnableWeakPtrThreadingAssertions::Yes)
+#endif
+    {
+        UNUSED_PARAM(shouldEnableAssertions);
+    }
+
     WeakRef(HashTableDeletedValueType) : m_impl(HashTableDeletedValue) { }
     WeakRef(HashTableEmptyValueType) : m_impl(HashTableEmptyValue) { }
 
     bool isHashTableDeletedValue() const { return m_impl.isHashTableDeletedValue(); }
     bool isHashTableEmptyValue() const { return m_impl.isHashTableEmptyValue(); }
+
+    Ref<WeakPtrImpl> releaseImpl() { return WTFMove(m_impl); }
 
     T* ptr() const
     {
@@ -69,6 +87,15 @@ public:
     {
         ASSERT(canSafelyBeUsed());
         return ptr();
+    }
+
+    EnableWeakPtrThreadingAssertions enableWeakPtrThreadingAssertions() const
+    {
+#if ASSERT_ENABLED
+        return m_shouldEnableAssertions ? EnableWeakPtrThreadingAssertions::Yes : EnableWeakPtrThreadingAssertions::No;
+#else
+        return EnableWeakPtrThreadingAssertions::No;
+#endif
     }
 
 private:
@@ -127,7 +154,7 @@ template<typename P, typename WeakPtrImpl> struct WeakRefHashTraits : SimpleClas
     static PeekType peek(P* value) { return value; }
 
     using TakeType = WeakPtr<P, WeakPtrImpl>;
-    static TakeType take(WeakRef<P, WeakPtrImpl>&& value) { return isEmptyValue(value) ? nullptr : WeakRef<P, WeakPtrImpl>(WTFMove(value)); }
+    static TakeType take(WeakRef<P, WeakPtrImpl>&& value) { return isEmptyValue(value) ? nullptr : WeakPtr<P, WeakPtrImpl>(WTFMove(value)); }
 };
 
 template<typename P, typename WeakPtrImpl> struct HashTraits<WeakRef<P, WeakPtrImpl>> : WeakRefHashTraits<P, WeakPtrImpl> { };
@@ -138,9 +165,139 @@ template<typename P, typename WeakPtrImpl> struct PtrHash<WeakRef<P, WeakPtrImpl
 
 template<typename P, typename WeakPtrImpl> struct DefaultHash<WeakRef<P, WeakPtrImpl>> : PtrHash<WeakRef<P, WeakPtrImpl>> { };
 
+DECLARE_COMPACT_ALLOCATOR_WITH_HEAP_IDENTIFIER(WeakPtrImplBase);
+
+template<typename Derived>
+class WeakPtrImplBase : public ThreadSafeRefCounted<Derived> {
+    WTF_MAKE_NONCOPYABLE(WeakPtrImplBase);
+    WTF_MAKE_FAST_COMPACT_ALLOCATED_WITH_HEAP_IDENTIFIER(WeakPtrImplBase);
+public:
+    ~WeakPtrImplBase() = default;
+
+    template<typename T> typename T::WeakValueType* get()
+    {
+        return static_cast<typename T::WeakValueType*>(m_ptr);
+    }
+
+    explicit operator bool() const { return m_ptr; }
+    void clear() { m_ptr = nullptr; }
+
+#if ASSERT_ENABLED
+    bool wasConstructedOnMainThread() const { return m_wasConstructedOnMainThread; }
+#endif
+
+    template<typename T>
+    explicit WeakPtrImplBase(T* ptr)
+        : m_ptr(static_cast<typename T::WeakValueType*>(ptr))
+#if ASSERT_ENABLED
+        , m_wasConstructedOnMainThread(isMainThread())
+#endif
+    {
+    }
+
+private:
+    void* m_ptr;
+#if ASSERT_ENABLED
+    bool m_wasConstructedOnMainThread;
+#endif
+};
+
+class DefaultWeakPtrImpl final : public WeakPtrImplBase<DefaultWeakPtrImpl> {
+public:
+    template<typename T>
+    explicit DefaultWeakPtrImpl(T* ptr) : WeakPtrImplBase<DefaultWeakPtrImpl>(ptr) { }
+};
+
+template<typename Derived>
+class WeakPtrImplBaseSingleThread {
+    WTF_MAKE_NONCOPYABLE(WeakPtrImplBaseSingleThread);
+    WTF_MAKE_FAST_COMPACT_ALLOCATED_WITH_HEAP_IDENTIFIER(WeakPtrImplBaseSingleThread);
+public:
+    ~WeakPtrImplBaseSingleThread() = default;
+
+    template<typename T> typename T::WeakValueType* get()
+    {
+        return static_cast<typename T::WeakValueType*>(m_ptr);
+    }
+
+    explicit operator bool() const { return m_ptr; }
+    void clear() { m_ptr = nullptr; }
+
+#if ASSERT_ENABLED
+    bool wasConstructedOnMainThread() const { return m_wasConstructedOnMainThread; }
+#endif
+
+    template<typename T>
+    explicit WeakPtrImplBaseSingleThread(T* ptr)
+        : m_ptr(static_cast<typename T::WeakValueType*>(ptr))
+#if ASSERT_ENABLED
+        , m_wasConstructedOnMainThread(isMainThread())
+#endif
+    {
+    }
+
+    uint32_t refCount() const { return m_refCount; }
+    void ref() const { ++m_refCount; }
+    void deref() const
+    {
+        uint32_t tempRefCount = m_refCount - 1;
+        if (!tempRefCount) {
+            delete this;
+            return;
+        }
+        m_refCount = tempRefCount;
+    }
+
+private:
+    mutable SingleThreadIntegralWrapper<uint32_t> m_refCount { 1 };
+    void* m_ptr;
+#if ASSERT_ENABLED
+    bool m_wasConstructedOnMainThread;
+#endif
+};
+
+class SingleThreadWeakPtrImpl final : public WeakPtrImplBaseSingleThread<SingleThreadWeakPtrImpl> {
+public:
+    template<typename T>
+    explicit SingleThreadWeakPtrImpl(T* ptr) : WeakPtrImplBaseSingleThread<SingleThreadWeakPtrImpl>(ptr) { }
+};
+
 template<typename T> using SingleThreadWeakRef = WeakRef<T, SingleThreadWeakPtrImpl>;
+
+template<typename ExpectedType, typename ArgType, typename WeakPtrImpl>
+inline bool is(WeakRef<ArgType, WeakPtrImpl>& source)
+{
+    return is<ExpectedType>(source.get());
+}
+
+template<typename ExpectedType, typename ArgType, typename WeakPtrImpl>
+inline bool is(const WeakRef<ArgType, WeakPtrImpl>& source)
+{
+    return is<ExpectedType>(source.get());
+}
+
+template<typename Target, typename Source, typename WeakPtrImpl>
+inline WeakRef<match_constness_t<Source, Target>, WeakPtrImpl> downcast(WeakRef<Source, WeakPtrImpl> source)
+{
+    static_assert(!std::is_same_v<Source, Target>, "Unnecessary cast to same type");
+    static_assert(std::is_base_of_v<Source, Target>, "Should be a downcast");
+    ASSERT_WITH_SECURITY_IMPLICATION(is<Target>(source));
+    return WeakRef<match_constness_t<Source, Target>, WeakPtrImpl> { static_reference_cast<match_constness_t<Source, Target>>(source.releaseImpl()), source.enableWeakPtrThreadingAssertions() };
+}
+
+template<typename Target, typename Source, typename WeakPtrImpl>
+inline WeakPtr<match_constness_t<Source, Target>, WeakPtrImpl> dynamicDowncast(WeakRef<Source, WeakPtrImpl> source)
+{
+    static_assert(!std::is_same_v<Source, Target>, "Unnecessary cast to same type");
+    static_assert(std::is_base_of_v<Source, Target>, "Should be a downcast");
+    if (!is<Target>(source))
+        return nullptr;
+    return WeakPtr<match_constness_t<Source, Target>, WeakPtrImpl> { static_reference_cast<match_constness_t<Source, Target>>(source.releaseImpl()), source.enableWeakPtrThreadingAssertions() };
+}
 
 } // namespace WTF
 
+using WTF::EnableWeakPtrThreadingAssertions;
+using WTF::SingleThreadWeakPtrImpl;
 using WTF::SingleThreadWeakRef;
 using WTF::WeakRef;
